Refactoring Plan 1: - Create a new app named `security`. - Refactor by splitting functionality related to `federal_web`. Migrate relevant components into the `account` and `security` apps. - Introduce individual `urls.py` files within each app for better modularity
