+++ id_broker/account/templates/federal_signin.html
@@ -60,13 +60,13 @@
 <body>
     <div class="container">
         <div class="login-box">
-            <button><a href="/account/login/">Sign In</a></button>
+            <button><a href="{{ baseURL }}/account/login/">Sign In</a></button>
         </div>
         
         <div class="login-options">
-            <button><a href="/oauth2/azure/auth">Sign in with Toastmasters</a></button>
-            <button><a href="/oauth2/google/auth">Sign in with Google</a></button>
-            <button><a href="/oauth2/line/auth">Sign in with Line</a></button>
+            <button><a href="{{ baseURL }}/oauth2/azure/auth">Sign in with Toastmasters</a></button>
+            <button><a href="{{ baseURL }}/oauth2/google/auth">Sign in with Google</a></button>
+            <button><a href="{{ baseURL }}/oauth2/line/auth">Sign in with Line</a></button>
         </div>
     </div>
 </body>
