+++ id_broker/account/urls.py
@@ -13,13 +13,25 @@ Including another URLconf
     1. Import the include() function: from django.urls import include, path
     2. Add a URL to urlpatterns:  path('blog/', include('blog.urls'))
 """
-from django.urls import path
+from django.urls import include, path
+from rest_framework.routers import DefaultRouter
 
 from . import views
+
+router = DefaultRouter()
+
+router.register(r"sign-up", views.IDRegister, basename="sign-up-account")
+router.register(r"profile", views.IDProfile, basename="user-profile")
+
 
 # Wire up our API using automatic URL routing.
 # Additionally, we include login URLs for the browsable API.
 urlpatterns = [
+    path(r"", include(router.urls)),
+    path(r"perform-confirmation/", views.perform_account_confirmation),
+    path(r"client-password-login/", views.client_password_login),
+    path(r"update-user-info/", views.UpdateUserInfoViews.as_view({"patch": "partial_update"})),
+    # Web page
     path(r"login-page", views.render_federal_signin_page),
     path(r"login-error-page", views.render_federal_oauth2_signin_error_page),
 ]
